Json read done
We should probably add it on it's own folders, test better the main folders and the ggplot2 is unreadable because of the huge amount of data making it laggy and unredable.

There's also an issue where I don't check if the data file exists, is just the reader that prompts an error.

But it's good enough to leave for the rest to fix, I am tired.